All SQL queries moved from collection.go
