+++ application.go
@@ -8,6 +8,7 @@ import (
 	"log"
 	"net/http"
 	"os"
+	"pebble-dev/rebblestore-api/db"
 	"time"
 
 	"github.com/gorilla/mux"
@@ -18,74 +19,14 @@ type PebbleAppList struct {
 	Apps []*PebbleApplication `json:"data"`
 }
 
-// RebbleApplication contains Pebble App information from the DB
-type RebbleApplication struct {
-	Id                 string        `json:"id"`
-	Name               string        `json:"title"`
-	Author             RebbleAuthor  `json:"author"`
-	Description        string        `json:"description"`
-	ThumbsUp           int           `json:"thumbs_up"`
-	Type               string        `json:"type"`
-	SupportedPlatforms []string      `json:"supported_platforms"`
-	Published          JSONTime      `json:"published_date"`
-	AppInfo            RebbleAppInfo `json:"appInfo"`
-	Assets             RebbleAssets  `json:"assets"`
-	DoomsdayBackup     bool          `json:"doomsday_backup"`
-}
-
 // RebbleTagList contains a list of tag. Used by getApi(id)
 type RebbleTagList struct {
-	Tags []RebbleCollection `json:"tags"`
+	Tags []db.RebbleCollection `json:"tags"`
 }
 
 // RebbleChangelog contains a list of version changes for an app
 type RebbleChangelog struct {
-	Versions []RebbleVersion `json:"versions"`
-}
-
-// RebbleVersion contains information about a specific version of an app
-type RebbleVersion struct {
-	Number      string   `json:"number"`
-	ReleaseDate JSONTime `json:"release_date"`
-	Description string   `json:"description"`
-}
-
-// RebbleAppInfo contains information about the app (pbw url, versioning, links, etc.)
-type RebbleAppInfo struct {
-	PbwUrl      string             `json:"pbwUrl"`
-	RebbleReady bool               `json:"rebbleReady"`
-	Tags        []RebbleCollection `json:"tags"`
-	Updated     JSONTime           `json:"updated"`
-	Version     string             `json:"version"`
-	SupportUrl  string             `json:"supportUrl"`
-	AuthorUrl   string             `json:"authorUrl"`
-	SourceUrl   string             `json:"sourceUrl"`
-}
-
-// RebbleAuthor describes the autor of a Rebble app (ID and name)
-type RebbleAuthor struct {
-	Id   int    `json:"id"`
-	Name string `json:"name"`
-}
-
-// RebbleAssets describes the list of assets of a Rebble app (banner, icon, screenshots)
-type RebbleAssets struct {
-	Banner      string                         `json:"appBanner"`
-	Icon        string                         `json:"appIcon"`
-	Screenshots *([]RebbleScreenshotsPlatform) `json:"screenshots"`
-}
-
-// RebbleScreenshotsPlatform contains a list of screenshots specific to some hardware (since each Pebble watch renders UI differently)
-type RebbleScreenshotsPlatform struct {
-	Platform    string   `json:"platform"`
-	Screenshots []string `json:"screenshots"`
-}
-
-// RebbleCollection describes the collection (category) of a Rebble application
-type RebbleCollection struct {
-	Id    string `json:"id"`
-	Name  string `json:"name"`
-	Color string `json:"color"`
+	Versions []db.RebbleVersion `json:"versions"`
 }
 
 // PebbleApplication is used by the parseApp() function. It matches directly the `{id}.json` format.
@@ -97,7 +38,7 @@ type PebbleApplication struct {
 	CategoryName       string                   `json:"category_name"`
 	CategoryColor      string                   `json:"category_color"`
 	Description        string                   `json:"description"`
-	Published          JSONTime                 `json:"published_date"`
+	Published          db.JSONTime              `json:"published_date"`
 	Release            PebbleApplicationRelease `json:"latest_release"`
 	Website            string                   `json:"website"`
 	Source             string                   `json:"source"`
@@ -113,17 +54,17 @@ type PebbleApplication struct {
 
 // PebbleApplicationRelease describes the `release` tag of a pebble JSON
 type PebbleApplicationRelease struct {
-	Id        string   `json:"id"`
-	PbwUrl    string   `json:"pbw_file"`
-	Published JSONTime `json:"published_date"`
-	Version   string   `json:"version"`
+	Id        string      `json:"id"`
+	PbwUrl    string      `json:"pbw_file"`
+	Published db.JSONTime `json:"published_date"`
+	Version   string      `json:"version"`
 }
 
 // PebbleVersion describes a version change
 type PebbleVersion struct {
-	Version   string   `json:"version"`
-	Published JSONTime `json:"published_date"`
-	Notes     string   `json:"release_notes"`
+	Version   string      `json:"version"`
+	Published db.JSONTime `json:"published_date"`
+	Notes     string      `json:"release_notes"`
 }
 
 // PebbleCompatibility describes the `compatibility` tag of a pebble JSON
@@ -187,7 +128,7 @@ func (pi *PebbleIcons) UnmarshalJSON(b []byte) error {
 	return json.Unmarshal(b, (*(map[string]string))(pi))
 }
 
-func parseApp(path string, authors *map[string]int, lastAuthorId *int, collections *map[string]RebbleCollection) (*RebbleApplication, *[]RebbleVersion, error) {
+func parseApp(path string, authors *map[string]int, lastAuthorId *int, collections *map[string]db.RebbleCollection) (*db.RebbleApplication, *[]db.RebbleVersion, error) {
 	f, err := ioutil.ReadFile(path)
 	if err != nil {
 		return nil, nil, err
@@ -212,16 +153,16 @@ func parseApp(path string, authors *map[string]int, lastAuthorId *int, collectio
 
 	// Create collection if it doesn't exist
 	if _, ok := (*collections)[data.Apps[0].CategoryId]; !ok {
-		(*collections)[data.Apps[0].CategoryId] = RebbleCollection{
+		(*collections)[data.Apps[0].CategoryId] = db.RebbleCollection{
 			Id:    data.Apps[0].CategoryId,
 			Name:  data.Apps[0].CategoryName,
 			Color: data.Apps[0].CategoryColor,
 		}
 	}
 
-	app := RebbleApplication{}
-	app.AppInfo.Tags = make([]RebbleCollection, 1)
-	screenshots := make(([]RebbleScreenshotsPlatform), 0)
+	app := db.RebbleApplication{}
+	app.AppInfo.Tags = make([]db.RebbleCollection, 1)
+	screenshots := make(([]db.RebbleScreenshotsPlatform), 0)
 	app.Assets.Screenshots = &screenshots
 
 	supportedPlatforms := make([]string, 0)
@@ -254,7 +195,7 @@ func parseApp(path string, authors *map[string]int, lastAuthorId *int, collectio
 	app.ThumbsUp = data.Apps[0].Hearts
 	app.Type = data.Apps[0].Type
 	app.SupportedPlatforms = supportedPlatforms
-	app.Author = RebbleAuthor{(*authors)[data.Apps[0].Author], data.Apps[0].Author}
+	app.Author = db.RebbleAuthor{(*authors)[data.Apps[0].Author], data.Apps[0].Author}
 	app.AppInfo.PbwUrl = data.Apps[0].Release.PbwUrl
 	app.AppInfo.RebbleReady = false
 	app.AppInfo.Updated = data.Apps[0].Release.Published
@@ -270,7 +211,7 @@ func parseApp(path string, authors *map[string]int, lastAuthorId *int, collectio
 	if icon, ok := data.Apps[0].Icons["48x48"]; ok {
 		app.Assets.Icon = icon
 	}
-	screenshots = append(*app.Assets.Screenshots, RebbleScreenshotsPlatform{data.Apps[0].ScreenshotHardware, make([]string, 0)})
+	screenshots = append(*app.Assets.Screenshots, db.RebbleScreenshotsPlatform{data.Apps[0].ScreenshotHardware, make([]string, 0)})
 	app.Assets.Screenshots = &screenshots
 	for _, screenshot := range data.Apps[0].Screenshots {
 		for _, s := range screenshot {
@@ -279,7 +220,7 @@ func parseApp(path string, authors *map[string]int, lastAuthorId *int, collectio
 	}
 	app.DoomsdayBackup = false
 
-	versions := make([]RebbleVersion, len(data.Apps[0].Changelog))
+	versions := make([]db.RebbleVersion, len(data.Apps[0].Changelog))
 	for i, pv := range data.Apps[0].Changelog {
 		versions[i].Number = pv.Version
 		versions[i].Description = pv.Notes
@@ -323,9 +264,9 @@ func RecurseFolder(w http.ResponseWriter, path string, f os.FileInfo, lvl int) {
 
 // AppsHandler lists all of the available applications from the backend DB.
 func AppsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (int, error) {
-	db := ctx.db
+	dbHandler := ctx.db
 
-	rows, err := db.Query(`
+	rows, err := dbHandler.Query(`
 			SELECT apps.name, authors.name
 			FROM apps
 			JOIN authors ON apps.author_id = authors.id
@@ -336,7 +277,7 @@ func AppsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (i
 	}
 
 	for rows.Next() {
-		item := RebbleApplication{}
+		item := db.RebbleApplication{}
 		err = rows.Scan(&item.Name, &item.Author.Name)
 		fmt.Fprintf(w, "Item: %s\n Author: %s\n\n", item.Name, item.Author.Name)
 	}
@@ -346,9 +287,9 @@ func AppsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (i
 
 // AppHandler returns a particular application from the backend DB as JSON
 func AppHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (int, error) {
-	db := ctx.db
+	dbHandler := ctx.db
 
-	rows, err := db.Query("SELECT apps.id, apps.name, apps.author_id, authors.name, apps.tag_ids, apps.description, apps.thumbs_up, apps.type, apps.supported_platforms, apps.published_date, apps.pbw_url, apps.rebble_ready, apps.updated, apps.version, apps.support_url, apps.author_url, apps.source_url, apps.screenshot_urls, apps.banner_url, apps.icon_url, apps.doomsday_backup FROM apps JOIN authors ON apps.author_id = authors.id WHERE apps.id=?", mux.Vars(r)["id"])
+	rows, err := dbHandler.Query("SELECT apps.id, apps.name, apps.author_id, authors.name, apps.tag_ids, apps.description, apps.thumbs_up, apps.type, apps.supported_platforms, apps.published_date, apps.pbw_url, apps.rebble_ready, apps.updated, apps.version, apps.support_url, apps.author_url, apps.source_url, apps.screenshot_urls, apps.banner_url, apps.icon_url, apps.doomsday_backup FROM apps JOIN authors ON apps.author_id = authors.id WHERE apps.id=?", mux.Vars(r)["id"])
 	if err != nil {
 		return http.StatusInternalServerError, err
 	}
@@ -358,13 +299,13 @@ func AppHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (in
 		return http.StatusNotFound, errors.New("No application with this ID")
 	}
 
-	app := RebbleApplication{}
+	app := db.RebbleApplication{}
 	var supportedPlatforms_b []byte
 	var t_published, t_updated int64
 	var tagIds_b []byte
 	var tagIds []string
 	var screenshots_b []byte
-	var screenshots *([]RebbleScreenshotsPlatform)
+	var screenshots *([]db.RebbleScreenshotsPlatform)
 	err = rows.Scan(&app.Id, &app.Name, &app.Author.Id, &app.Author.Name, &tagIds_b, &app.Description, &app.ThumbsUp, &app.Type, &supportedPlatforms_b, &t_published, &app.AppInfo.PbwUrl, &app.AppInfo.RebbleReady, &t_updated, &app.AppInfo.Version, &app.AppInfo.SupportUrl, &app.AppInfo.AuthorUrl, &app.AppInfo.SourceUrl, &screenshots_b, &app.Assets.Banner, &app.Assets.Icon, &app.DoomsdayBackup)
 	if err != nil {
 		return http.StatusInternalServerError, err
@@ -373,12 +314,12 @@ func AppHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (in
 	app.Published.Time = time.Unix(0, t_published)
 	app.AppInfo.Updated.Time = time.Unix(0, t_updated)
 	json.Unmarshal(tagIds_b, &tagIds)
-	app.AppInfo.Tags = make([]RebbleCollection, len(tagIds))
+	app.AppInfo.Tags = make([]db.RebbleCollection, len(tagIds))
 	json.Unmarshal(screenshots_b, &screenshots)
 	app.Assets.Screenshots = screenshots
 
 	for i, tagId := range tagIds {
-		rows, err := db.Query("SELECT id, name, color FROM collections WHERE id=?", tagId)
+		rows, err := dbHandler.Query("SELECT id, name, color FROM collections WHERE id=?", tagId)
 		if err != nil {
 			return http.StatusInternalServerError, err
 		}
@@ -403,9 +344,9 @@ func AppHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (in
 
 // TagsHandler returns the list of tags of a particular appliction as JSON
 func TagsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (int, error) {
-	db := ctx.db
+	dbHandler := ctx.db
 
-	rows, err := db.Query("SELECT apps.tag_ids FROM apps")
+	rows, err := dbHandler.Query("SELECT apps.tag_ids FROM apps")
 	if err != nil {
 		return http.StatusInternalServerError, err
 	}
@@ -422,10 +363,10 @@ func TagsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (i
 	}
 	json.Unmarshal(tagIds_b, &tagIds)
 	tagList := RebbleTagList{}
-	tagList.Tags = make([]RebbleCollection, len(tagIds))
+	tagList.Tags = make([]db.RebbleCollection, len(tagIds))
 
 	for i, tagId := range tagIds {
-		rows, err := db.Query("SELECT id, name, color FROM collections WHERE id=?", tagId)
+		rows, err := dbHandler.Query("SELECT id, name, color FROM collections WHERE id=?", tagId)
 		if err != nil {
 			return http.StatusInternalServerError, err
 		}
@@ -451,9 +392,9 @@ func TagsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (i
 
 // VersionsHandler returns the server version
 func VersionsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request) (int, error) {
-	db := ctx.db
+	dbHandler := ctx.db
 
-	rows, err := db.Query("SELECT apps.versions FROM apps")
+	rows, err := dbHandler.Query("SELECT apps.versions FROM apps")
 	if err != nil {
 		return http.StatusInternalServerError, err
 	}
@@ -463,7 +404,7 @@ func VersionsHandler(ctx *handlerContext, w http.ResponseWriter, r *http.Request
 	}
 
 	var versions_b []byte
-	var versions []RebbleVersion
+	var versions []db.RebbleVersion
 	err = rows.Scan(&versions_b)
 	if err != nil {
 		return http.StatusInternalServerError, err
